Move the Serializer api and SerializerRegistery impl to "-runtime"
While we don't necessarily want to allow other SerializerRegistry
implementations, the logic belongs in our runtime module, not
the library.
