+++ conjure-java-undertow-runtime/src/test/java/com/palantir/conjure/java/undertow/demo/MyServiceUndertowHandlerTest.java
@@ -21,8 +21,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.google.common.collect.ImmutableList;
 import com.google.common.reflect.TypeToken;
 import com.palantir.conjure.java.undertow.lib.Routable;
-import com.palantir.conjure.java.undertow.lib.Serializer;
 import com.palantir.conjure.java.undertow.runtime.ConjureHandler;
+import com.palantir.conjure.java.undertow.runtime.Serializer;
 import com.palantir.conjure.java.undertow.runtime.Serializers;
 import com.palantir.logsafe.exceptions.SafeRuntimeException;
 import com.palantir.tracing.Tracers;
