+++ conjure-java-undertow-runtime/src/test/java/com/palantir/conjure/java/undertow/runtime/SerializerRegistryTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.conjure.java.undertow.lib;
+package com.palantir.conjure.java.undertow.runtime;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -37,7 +37,7 @@ public class SerializerRegistryTest {
 
         HttpServerExchange exchange = HttpServerExchanges.createStub();
         exchange.getRequestHeaders().put(Headers.CONTENT_TYPE, "text/plain");
-        SerializerRegistry serializers = new SerializerRegistry(json, plain);
+        ConjureSerializerRegistry serializers = new ConjureSerializerRegistry(json, plain);
         Serializer serializer = serializers.getRequestDeserializer(exchange);
         assertThat(serializer).isSameAs(plain);
     }
@@ -45,7 +45,7 @@ public class SerializerRegistryTest {
     @Test
     public void testRequestNoContentType() {
         HttpServerExchange exchange = HttpServerExchanges.createStub();
-        SerializerRegistry serializers = new SerializerRegistry(new StubSerializer("application/json"));
+        ConjureSerializerRegistry serializers = new ConjureSerializerRegistry(new StubSerializer("application/json"));
         assertThatThrownBy(() -> serializers.getRequestDeserializer(exchange))
                 .isInstanceOf(SafeIllegalArgumentException.class)
                 .hasMessageContaining("Request is missing Content-Type header");
@@ -55,7 +55,7 @@ public class SerializerRegistryTest {
     public void testUnsupportedRequestContentType() {
         HttpServerExchange exchange = HttpServerExchanges.createStub();
         exchange.getRequestHeaders().put(Headers.CONTENT_TYPE, "application/unknown");
-        SerializerRegistry serializers = new SerializerRegistry(new StubSerializer("application/json"));
+        ConjureSerializerRegistry serializers = new ConjureSerializerRegistry(new StubSerializer("application/json"));
         assertThatThrownBy(() -> serializers.getRequestDeserializer(exchange))
                 .isInstanceOf(SafeIllegalArgumentException.class)
                 .hasMessageContaining("Unsupported Content-Type");
@@ -68,7 +68,7 @@ public class SerializerRegistryTest {
 
         HttpServerExchange exchange = HttpServerExchanges.createStub();
         exchange.getRequestHeaders().put(Headers.ACCEPT, "text/plain");
-        SerializerRegistry serializers = new SerializerRegistry(json, plain);
+        ConjureSerializerRegistry serializers = new ConjureSerializerRegistry(json, plain);
         Serializer serializer = serializers.getResponseSerializer(exchange);
         assertThat(serializer).isSameAs(plain);
     }
@@ -79,7 +79,7 @@ public class SerializerRegistryTest {
         Serializer plain = new StubSerializer("text/plain");
 
         HttpServerExchange exchange = HttpServerExchanges.createStub();
-        SerializerRegistry serializers = new SerializerRegistry(json, plain);
+        ConjureSerializerRegistry serializers = new ConjureSerializerRegistry(json, plain);
         Serializer serializer = serializers.getResponseSerializer(exchange);
         assertThat(serializer).isSameAs(json);
     }
@@ -91,7 +91,7 @@ public class SerializerRegistryTest {
 
         HttpServerExchange exchange = HttpServerExchanges.createStub();
         exchange.getRequestHeaders().put(Headers.ACCEPT, "application/unknown");
-        SerializerRegistry serializers = new SerializerRegistry(json, plain);
+        ConjureSerializerRegistry serializers = new ConjureSerializerRegistry(json, plain);
         Serializer serializer = serializers.getResponseSerializer(exchange);
         assertThat(serializer).isSameAs(json);
     }
